remove public declared
From section 9.4 of the Java Language Specification:
Every method declaration in the body of an interface is implicitly public (§6.6).
Every method declaration in the body of an interface is implicitly abstract, so its body is always represented by a semicolon, not a block.
It is permitted, but discouraged as a matter of style, to redundantly specify the public and/or abstract modifier for a method declared in an interface.
So it‘s no use to Modifier 'public' is redundant for interface methods. remove it to make code clear.